+++ example/integration_test/migration_test.dart
@@ -8,7 +8,6 @@ import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as path;
 import 'package:background_downloader/src/localstore/localstore.dart';
 
-
 const tasksPath = 'backgroundDownloaderTaskRecords';
 const databaseMetadataPath = 'backgroundDownloaderDatabase';