+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## 7.3.1
+
+Improved [DownloadProgressIndicator](https://pub.dev/documentation/background_downloader/latest/background_downloader/DownloadProgressIndicator-class.html) widget:
+* In collapsed state, now shows progress as 'n' files finished out of 'total' started (and progress as that fraction)
+* Option to force collapsed state always by setting `maxExpandable` to 0. When set to 1, the indicator collapses only when the second download starts. When set greater than 1, the indicator expands to show multiple simultaneous downloads.
+
+Added usage examples upfront in the readme
+
 ## 7.3.0
 
-Added `DownloadProgressIndicator` widget and modified the example app to show how to wire it up.
+Added [DownloadProgressIndicator](https://pub.dev/documentation/background_downloader/latest/background_downloader/DownloadProgressIndicator-class.html) widget and modified the example app to show how to wire it up.
 
 The widget is configurable (e.g. pause and cancel buttons) and can show multiple downloads simultaneously in either an expanded
 or collapsed mode.
@@ -11,7 +19,7 @@ Fixed bugs.
 
 ## 7.2.0
 
-Added option to use a different persistent storage solution than the one provided by default. The downloader stores a few things in persistent storage, and uses a modified version of the [localstore](https://pub.dev/packages/localstore) package by default. To use a different persistent storage solution, create a class that implements the `PersistentStorage` interface, and initialize the downloader by calling `FileDownloader(persistentStorage: yourStorageClass())` as the first use of the `FileDownloader`.
+Added option to use a different persistent storage solution than the one provided by default. The downloader stores a few things in persistent storage, and uses a modified version of the [localstore](https://pub.dev/packages/localstore) package by default. To use a different persistent storage solution, create a class that implements the [PersistentStorage](https://pub.dev/documentation/background_downloader/latest/background_downloader/PersistentStorage-class.html) interface, and initialize the downloader by calling `FileDownloader(persistentStorage: yourStorageClass())` as the first use of the `FileDownloader`.
 
 A simple example is included in the example app (using the [sqflite](https://pub.dev/packages/sqflite) package).